requirement.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==1.11.10
django-braces==1.12.0
django-filter==1.1.0
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.7.7
ecdsa==0.13
enum34==1.1.6
future==0.16.0
httpie==0.9.9
idna==2.6
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
Markdown==2.6.11
MarkupSafe==1.0
oauthlib==2.0.6
opentok==2.6.0
Pillow==5.0.0
postgres==2.2.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycryptodome==3.4.11
Pygments==2.2.0
PyJWT==1.5.3
python-dateutil==2.6.1
python-jose==2.0.2
python-openid==2.2.5
pytz==2018.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sorl-thumbnail==12.4.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
